create display_utils
