+++ ex36 - without is_game_continue.py
@@ -2,6 +2,8 @@ from sys import exit, stdout
 import time
 import webbrowser
 
+from display_utils import display_text
+
 # Global constants
 MENU_DELAY: float = 0.5
 
@@ -221,22 +223,6 @@ def good_ending():
 	display_text("A new journey awaits you ahead.")
 	display_text("TO BE CONTINUED...")
 	exit(0)
-
-def display_text(*args: any, end: str = '\n') -> None:
-	"""
-	Prints given texts, character by character with a delay between each character.
-	"""
-	DELAY: float = 0.05
-	len_args: int = len(args)
-	for i in range(len_args):
-		text: str = str(args[i])
-		for char in text:
-			print(char, end='')
-			stdout.flush() 
-			time.sleep(DELAY)
-		if (i != len_args - 1):
-			print(' ', end='')
-	print(end=end)
 
 # Begin game
 start()
